pmtparse: fix sign compare warning in getDefaultAudioPid
This commit fixes the following warning:
dvb/pmtparse.cpp: In member function ‘virtual RESULT eDVBPMTParser::eStreamData::getDefaultAudioPid(int&) const’:
dvb/pmtparse.cpp:612:26: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  if (audioStreams.size() > defaultAudioPid)
      ~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~

+++ lib/dvb/pmtparse.cpp
@@ -609,7 +609,7 @@ RESULT eDVBPMTParser::eStreamData::getCaIds(std::vector<int> &caids, std::vector
 
 RESULT eDVBPMTParser::eStreamData::getDefaultAudioPid(int &result) const
 {
-	if (audioStreams.size() > defaultAudioPid)
+	if (audioStreams.size() > static_cast<size_t>(defaultAudioPid))
 	{
 		result = audioStreams[defaultAudioPid];
 	}
